pyproject.toml
[project]
name = "traffic-workshop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faker>=37.11.0",
    "google-adk>=1.16.0",
    "toolbox-core>=0.5.2",
]
